UI: Set bigger default window size
This improves the situation when the user enables detailed view
on data with a lot of child functions. Previously the legend would
take up practically all the space and hide the actual graph.

However, it is not a good long-term solution as not all users will
have such a big screen. In addition, larger data-sets might very well
require the window to be even bigger to be useful.